requirements.txt
redis-om
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
cryptography<40
requests